Fixed a bunch of bugs. Changed it so updates not collusions went to Fixed Update. Added default icon files.

+++ Defender the Remake/TheBackground.h
@@ -9,7 +9,7 @@ public:
 	TheBackground();
 	virtual ~TheBackground();
 
-	void SetLandPartsModel(Model landPart, Model radarLandPart, int index);
+	void SetLandPartsModel(Model landPart, Model radarLandPart, size_t index);
 	void SetUIBackfaceModel(Model model);
 	void SetRadarModel(Model horizontal, Model side);
 	void SetStarModel(Model model);
